chore: convert test code to TypeScript
This is good for checking our own type definitions make sense. Some types were updated based on their usage in the tests.
