+++ test/unit/nodeunitheadless.js
@@ -1,6 +1,2 @@
-var TWEEN_uncompressed = require('../../dist/tween.cjs.js')
-var getTests = require('./tests')
-
-module.exports = {
-	tween: getTests(TWEEN_uncompressed),
-}
+const {tests} = require('../../.tmp/tests.cjs.js')
+module.exports = {tween: tests}
